Changed tune-lowlevel slightly so that the library is always rebuilt afresh
This cures the oddity of mul8k showing absurd timings because of source
files dependencies not being properly caught by the building system while
doing the tuning.

The time it takes to build the library itself is relatively negligible
compared to the time we invest in tuning, so it makes sense to proceed
like this.

git-svn-id: svn://scm.gforge.inria.fr/svnroot/gf2x/trunk@164 e5c1114b-a573-4582-9dac-f72f410959ce